Zero out the time when using start and end dates
Also, it doesn't make sense to filter out loans by the end date, since
if it has an end date, it will have been returned. Because of that, we
could get Loan objects in the form of CalendarItem objects that don't
fall within the range of the calendar, so we have to have a special
check to see if we should add them to the calendar.