Always show Items first in the autocomplete
Items will be entered via a barcode, so that is the only field we match
on. And we want the Item to be the first in the autocomplete list, so
when the tab key is used, the Item is immediately selected, instead of a
Bib.

Now we can get rid of autocomplete for just bibs, and replace it with a
single autocomplete that shows Items and then Bibs